requirements.txt
streamlit
ibm-watson
ibm-cloud-sdk-core
python-dotenv
pip install -r requirements.txt
